Changed Defen to static library and improved premake files.

+++ Defen/src/Defen/ImGui/ImGuiLayer.cpp
@@ -5,8 +5,6 @@
 #include "Defen/Application.h"
 #include <backends/imgui_impl_glfw.h>
 #include <backends/imgui_impl_opengl3.h>
-#include <backends/imgui_impl_opengl3.cpp>
-#include <backends/imgui_impl_glfw.cpp>
 #include "GLFW/glfw3.h"
 
 namespace Defen
@@ -71,7 +69,7 @@ namespace Defen
 	{
 		ImGuiIO& io = ImGui::GetIO();
 		Application& app = Application::Get();
-		io.DisplaySize = ImVec2(app.GetWindow().GetWidth(), app.GetWindow().GetHeight());
+		io.DisplaySize = ImVec2((float)app.GetWindow().GetWidth(), (float)app.GetWindow().GetHeight());
 
 		ImGui::Render();
 		ImGui_ImplOpenGL3_RenderDrawData(ImGui::GetDrawData());
